server: don't truncate SSIZE_MAX when guarding file sizes
dqlite_server_start guards the node-info and node-store file sizes
with `full_size > (off_t)SSIZE_MAX`. Casting the limit *down* to off_t
truncates whenever off_t is narrower than ssize_t: on Windows/UCRT
(off_t is a 32-bit long, ssize_t is 64-bit) SSIZE_MAX becomes -1, so
the guard trips for every file size -- even 0 -- and
dqlite_server_start fails unconditionally. Cast the operands up to
int64_t instead.

Linux behaviour is unchanged: with _FILE_OFFSET_BITS=64 (set by the
build, matching AC_SYS_LARGEFILE) off_t is a 64-bit signed type, so
both operands hold exactly the same values as before. On a
hypothetical ILP32-with-LFS target (64-bit off_t, 32-bit ssize_t) the
old code already widened correctly and the new code is again
value-identical. The LHS cast too keeps clang's default-on
-Wtautological-constant-out-of-range-compare quiet where off_t is
32-bit.

This is the root cause that compat/win's _OFF_T_DEFINED/64-bit-off_t
hijack was papering over; that hijack is removed in the next commit.

Verified on Windows: integration server suite 8/8 (the suite that
regressed when this guard misfired), plus unit-test 319/319.

+++ src/server.c
@@ -1932,7 +1932,10 @@ int dqlite_server_start(dqlite_server *server)
 
 	full_size = lseek(info_fd, 0, SEEK_END);
 	dqlite_assert(full_size >= 0);
-	if (full_size > (off_t)SSIZE_MAX) {
+	/* Compare in a 64-bit type: casting SSIZE_MAX *down* to off_t instead
+	 * would truncate (to -1) on platforms where off_t is narrower than
+	 * ssize_t, making this guard trip for every file size. */
+	if ((int64_t)full_size > (int64_t)SSIZE_MAX) {
 		goto err_after_open_store;
 	}
 	size = (ssize_t)full_size;
@@ -1957,7 +1960,8 @@ int dqlite_server_start(dqlite_server *server)
 
 	full_size = lseek(store_fd, 0, SEEK_END);
 	dqlite_assert(full_size >= 0);
-	if (full_size > (off_t)SSIZE_MAX) {
+	/* See the 64-bit comparison rationale above. */
+	if ((int64_t)full_size > (int64_t)SSIZE_MAX) {
 		goto err_after_open_store;
 	}
 	size = (ssize_t)full_size;
